chore: no more v2 for operation related services
