+++ bc_obps/registration/api/_operations/_operation_id/_registration/opted_in_operation.py
@@ -7,7 +7,7 @@ from registration.schema import (
     OptedInOperationDetailOut,
     Message
 )
-from service.operation_service_v2 import OperationServiceV2
+from service.operation_service import OperationService
 from registration.constants import OPERATION_TAGS
 from common.permissions import authorize
 from common.api.utils import get_current_user_guid
@@ -26,7 +26,7 @@ from registration.api.router import router
 def operation_registration_get_opted_in_operation_detail(
     request: HttpRequest, operation_id: UUID
 ) -> Tuple[Literal[200], Optional[OptedInOperationDetail]]:
-    return 200, OperationServiceV2.get_opted_in_operation_detail(get_current_user_guid(request), operation_id)
+    return 200, OperationService.get_opted_in_operation_detail(get_current_user_guid(request), operation_id)
 
 
 @router.put(
@@ -40,6 +40,6 @@ def operation_registration_get_opted_in_operation_detail(
 def operation_registration_update_opted_in_operation_detail(
     request: HttpRequest, operation_id: UUID, payload: OptedInOperationDetailIn
 ) -> Tuple[Literal[200, 400], OptedInOperationDetail]:
-    return 200, OperationServiceV2.update_opted_in_operation_detail(
+    return 200, OperationService.update_opted_in_operation_detail(
         get_current_user_guid(request), operation_id, payload
     )
