+++ bc_obps/registration/api/_operations/current.py
@@ -4,7 +4,7 @@ from registration.models.operation import Operation
 from common.api.utils import get_current_user_guid
 from registration.schema import OperationCurrentOut, Message
 from service.error_service.custom_codes_4xx import custom_codes_4xx
-from service.operation_service_v2 import OperationServiceV2
+from service.operation_service import OperationService
 from registration.api.router import router
 from common.permissions import authorize
 from django.db.models import QuerySet
@@ -21,4 +21,4 @@ from registration.constants import OPERATION_TAGS
     The endpoint ensures that only authorized industry users can get unregistered operations belonging to their operator. Unauthorized access attempts raise an error.""",
 )
 def list_current_users_operations(request: HttpRequest) -> Tuple[Literal[200], QuerySet[Operation]]:
-    return 200, OperationServiceV2.list_current_users_unregistered_operations(get_current_user_guid(request))
+    return 200, OperationService.list_current_users_unregistered_operations(get_current_user_guid(request))
