+++ bc_obps/service/data_access_service/opted_in_operation_detail_service.py
@@ -1,9 +1,9 @@
 from uuid import UUID
-from service.operation_service import OperationService
 from registration.schema import OptedInOperationDetailIn
 from registration.models.opted_in_operation_detail import OptedInOperationDetail
 from registration.models.operation import Operation
 from registration.utils import update_model_instance
+from service.operation_service import OperationService
 
 
 class OptedInOperationDataAccessService:
