+++ bc_obps/service/transfer_event_service.py
@@ -26,7 +26,7 @@ from service.data_access_service.user_service import UserDataAccessService
 from service.facility_designated_operation_timeline_service import FacilityDesignatedOperationTimelineService
 from service.facility_service import FacilityService
 from service.operation_designated_operator_timeline_service import OperationDesignatedOperatorTimelineService
-from service.operation_service_v2 import OperationServiceV2
+from service.operation_service import OperationService
 
 logger = logging.getLogger(__name__)
 
@@ -289,7 +289,7 @@ class TransferEventService:
         )
 
         # update the operation's operator
-        OperationServiceV2.update_operator(user_guid, event.operation, event.to_operator.id)  # type: ignore # we are sure that operation is not None
+        OperationService.update_operator(user_guid, event.operation, event.to_operator.id)  # type: ignore # we are sure that operation is not None
 
     @classmethod
     def get_if_authorized(cls, user_guid: UUID, transfer_id: UUID) -> TransferEvent:
